+++ apps/web/app/routes.ts
@@ -82,6 +82,10 @@ export default [
                 'oasis-bonus-finder',
                 '(game)/(village-slug)/(oasis-bonus-finder)/page.tsx',
               ),
+              route(
+                'oasis-animal-finder',
+                '(game)/(village-slug)/(oasis-animal-finder)/page.tsx',
+              ),
               route('events', '(game)/(village-slug)/(events)/page.tsx'),
               ...prefix('players', [
                 index('(game)/(village-slug)/(players)/page.tsx'),